[#15033] Move PrivateIndexingConfiguration to internal package in query
