+++ query/src/test/java/org/infinispan/query/reindex/ReindexPressureTest.java
@@ -8,12 +8,12 @@ import java.util.Map;
 
 import org.infinispan.configuration.cache.ConfigurationBuilder;
 import org.infinispan.configuration.cache.IndexingMode;
-import org.infinispan.configuration.cache.PrivateIndexingConfigurationBuilder;
 import org.infinispan.manager.EmbeddedCacheManager;
 import org.infinispan.query.Indexer;
 import org.infinispan.query.Search;
 import org.infinispan.query.core.stats.IndexInfo;
 import org.infinispan.query.core.stats.IndexStatistics;
+import org.infinispan.query.impl.config.PrivateIndexingConfigurationBuilder;
 import org.infinispan.query.model.Game;
 import org.infinispan.test.SingleCacheManagerTest;
 import org.infinispan.test.fwk.TestCacheManagerFactory;
